refactor(api): #3323 split payment flat specific code and rework data bretagne (#3325)
* refactor(api): fonjep - create dbo from entities

* refactor(api): save null instead of empty string in fonjep data

* refactor(api): make only fonjep code mandatory in types

* refactor(api): split payment flat specific code and rework data bretagne
