+++ packages/api/src/dataProviders/db/paymentFlat/paymentFlat.port.test.ts
@@ -1,8 +1,8 @@
-import { PAYMENT_FLAT_ENTITY } from "./__fixtures__/paymentFlatEntity.fixture";
+import { PAYMENT_FLAT_ENTITY } from "../../../modules/paymentFlat/__fixtures__/paymentFlatEntity.fixture";
+import PaymentFlatAdapter from "../../../modules/paymentFlat/paymentFlatAdapter";
 import { PAYMENT_FLAT_DBO } from "./__fixtures__/paymentFlatDbo.fixture";
 import paymentFlatPort from "./paymentFlat.port";
 import { ObjectId } from "mongodb";
-import PaymentsFlatAdapter from "./PaymentFlat.adapter";
 const mockDeleteMany = jest.fn();
 const mockInsertOne = jest.fn();
 const mockUpdateOne = jest.fn();
@@ -16,20 +16,23 @@ jest.mock("../../../shared/MongoConnection", () => ({
 }));
 
 describe("PaymentFlat Port", () => {
+    beforeEach(() => {
+        jest.spyOn(PaymentFlatAdapter, "toDbo").mockReturnValue(PAYMENT_FLAT_DBO);
+    });
+
     describe("insertOne()", () => {
         it("should call insertOne with the correct arguments", async () => {
             await paymentFlatPort.insertOne(PAYMENT_FLAT_ENTITY);
-            expect(mockInsertOne).toHaveBeenCalledWith({ ...PAYMENT_FLAT_DBO, _id: expect.any(ObjectId) });
+            expect(mockInsertOne).toHaveBeenCalledWith(PAYMENT_FLAT_DBO);
         });
     });
 
     describe("upsertOne()", () => {
-        it("should call upsertOne with the correct arguments", async () => {
+        it("should call toDbo()", async () => {
             await paymentFlatPort.upsertOne(PAYMENT_FLAT_ENTITY);
-            const { _id, ...DboWithoutId } = PAYMENT_FLAT_DBO;
             expect(mockUpdateOne).toHaveBeenCalledWith(
-                { uniqueId: PAYMENT_FLAT_ENTITY.uniqueId },
-                { $set: DboWithoutId },
+                { uniqueId: PAYMENT_FLAT_DBO.uniqueId },
+                { $set: PAYMENT_FLAT_DBO },
                 { upsert: true },
             );
         });
@@ -44,10 +47,8 @@ describe("PaymentFlat Port", () => {
 
     describe("cursorFindChorusOnly()", () => {
         let mockCursorFind: jest.SpyInstance;
-        let mockDboToEntity: jest.SpyInstance;
         beforeAll(() => {
             mockCursorFind = jest.spyOn(paymentFlatPort, "cursorFind").mockImplementation(jest.fn());
-            mockDboToEntity = jest.spyOn(PaymentsFlatAdapter, "dboToEntity").mockImplementation(jest.fn());
         });
 
         afterAll(() => {
