+++ packages/api/src/dataProviders/db/paymentFlat/paymentFlat.port.ts
@@ -3,10 +3,11 @@ import PaymentFlatEntity from "../../../entities/PaymentFlatEntity";
 import Siren from "../../../valueObjects/Siren";
 import Siret from "../../../valueObjects/Siret";
 import { DefaultObject } from "../../../@types";
+import PaymentFlatAdapter from "../../../modules/paymentFlat/paymentFlatAdapter";
 import PaymentFlatDbo from "./PaymentFlatDbo";
-import PaymentFlatAdapter from "./PaymentFlat.adapter";
 
-export class PaymentFlatPort extends MongoPort<PaymentFlatDbo> {
+// we omit _id because it is generated by MongoDB
+export class PaymentFlatPort extends MongoPort<Omit<PaymentFlatDbo, "_id">> {
     collectionName = "payments-flat";
 
     public async createIndexes(): Promise<void> {
@@ -26,8 +27,7 @@ export class PaymentFlatPort extends MongoPort<PaymentFlatDbo> {
 
     public upsertOne(entity: PaymentFlatEntity) {
         const updateDbo = PaymentFlatAdapter.toDbo(entity);
-        const { _id, ...DboWithoutId } = updateDbo;
-        return this.collection.updateOne({ uniqueId: updateDbo.uniqueId }, { $set: DboWithoutId }, { upsert: true });
+        return this.collection.updateOne({ uniqueId: updateDbo.uniqueId }, { $set: updateDbo }, { upsert: true });
     }
 
     public upsertMany(bulk) {
