+++ packages/api/src/interfaces/cli/PaymentFlat.cli.test.ts
@@ -1,5 +1,7 @@
+import paymentFlatChorusService from "../../modules/paymentFlat/paymentFlat.chorus.service";
+import paymentFlatService from "../../modules/paymentFlat/paymentFlat.service";
 import PaymentFlatCli from "./PaymentFlat.cli";
-import paymentsFlatService from "../../modules/paymentFlat/paymentFlat.service";
+jest.mock("../../modules/paymentFlat/paymentFlat.chorus.service");
 jest.mock("../../modules/paymentFlat/paymentFlat.service");
 
 describe("PaymentFlat Cli", () => {
@@ -13,31 +15,32 @@ describe("PaymentFlat Cli", () => {
         it("should call updatePaymentsFlatCollection with the given exerciceBudgetaire", async () => {
             const exerciceBudgetaire = 2022;
             await paymentFlatCli.resyncExercice(exerciceBudgetaire);
-            expect(paymentsFlatService.updatePaymentsFlatCollection).toHaveBeenCalledWith(exerciceBudgetaire);
+            expect(paymentFlatChorusService.updatePaymentsFlatCollection).toHaveBeenCalledWith(exerciceBudgetaire);
         });
     });
 
     describe("init", () => {
-        const mockIsCollectionInitialized = jest.spyOn(paymentsFlatService, "isCollectionInitialized");
+        jest.spyOn(global, "setInterval").mockImplementation(jest.fn());
+        const mockIsCollectionInitialized = jest.spyOn(paymentFlatService, "isCollectionInitialized");
         beforeAll(() => {
             mockIsCollectionInitialized.mockResolvedValue(false);
         });
 
         it("calls isCollectionInitialized", async () => {
-            await paymentFlatCli.init();
-            expect(paymentsFlatService.isCollectionInitialized).toHaveBeenCalledTimes(1);
+            await paymentFlatCli.initChorus();
+            expect(paymentFlatService.isCollectionInitialized).toHaveBeenCalledTimes(1);
         });
 
         it("throws an error if collection has already been initialized", () => {
             mockIsCollectionInitialized.mockResolvedValueOnce(true);
-            expect(async () => await paymentFlatCli.init()).rejects.toThrowError(
+            expect(async () => await paymentFlatCli.initChorus()).rejects.toThrowError(
                 "DB already initialized, used resyncExercice instead",
             );
         });
 
         it("calls service init method", async () => {
-            await paymentFlatCli.init();
-            expect(paymentsFlatService.init).toHaveBeenCalledTimes(1);
+            await paymentFlatCli.initChorus();
+            expect(paymentFlatChorusService.init).toHaveBeenCalledTimes(1);
         });
     });
 });
