+++ packages/api/src/interfaces/cli/PaymentFlat.cli.ts
@@ -1,4 +1,5 @@
 import { StaticImplements } from "../../decorators/staticImplements.decorator";
+import paymentFlatChorusService from "../../modules/paymentFlat/paymentFlat.chorus.service";
 import { CliStaticInterface } from "../../@types";
 
 import paymentsFlatService from "../../modules/paymentFlat/paymentFlat.service";
@@ -13,18 +14,18 @@ export default class PaymentFlatCli extends CliController {
         if (!exerciceBudgetaire) throw new Error("Exercice budgetaire is required");
 
         this.logger.logIC(`Resync payment flat collection for exercice ${exerciceBudgetaire}`);
-        await paymentsFlatService.updatePaymentsFlatCollection(exerciceBudgetaire);
+        await paymentFlatChorusService.updatePaymentsFlatCollection(exerciceBudgetaire);
         clearInterval(ticTacInterval);
     }
 
     // should only be used once, then sync with resyncExercise
-    async init() {
+    async initChorus() {
         if (await paymentsFlatService.isCollectionInitialized())
             throw new Error("DB already initialized, used resyncExercice instead");
 
         this.logger.logIC("Create all payment flat collection");
         const ticTacInterval = setInterval(() => console.log("TIC"), 60000);
-        await paymentsFlatService.init();
+        await paymentFlatChorusService.init();
         clearInterval(ticTacInterval);
     }
 }
