+++ packages/api/src/interfaces/http/DataViz.http.test.ts
@@ -13,6 +13,7 @@ describe("DataVizHttp", () => {
             amountsVsProgramRegionService,
             "getAmountsVsProgramRegionData",
         );
+
         beforeAll(() => {
             getAmountsVsProgramRegionDataSpy.mockResolvedValue(AmountsVsProgramRegionDtoList);
         });
