+++ packages/api/src/modules/dataViz/amountsVsProgramRegion/amountsVsProgramRegion.adapter.ts
@@ -1,13 +1,12 @@
 import PaymentFlatEntity from "../../../entities/PaymentFlatEntity";
+import { AmountsVsProgramRegionDbo } from "./entitiyAndDbo/amountsVsProgramRegion.dbo";
 import AmountsVsProgramRegionEntity from "./entitiyAndDbo/amountsVsProgramRegion.entity";
 
 export default class AmountsVsProgramRegionAdapter {
     static toNotAggregatedEntity(paymentFlatEntity: PaymentFlatEntity): AmountsVsProgramRegionEntity {
         const amountsVsProgramRegionEntity = {
             exerciceBudgetaire: paymentFlatEntity.exerciceBudgetaire,
-            regionAttachementComptable: paymentFlatEntity.regionAttachementComptable
-                ? paymentFlatEntity.regionAttachementComptable
-                : "Non renseigné",
+            regionAttachementComptable: paymentFlatEntity.regionAttachementComptable,
             programme: paymentFlatEntity.programName
                 ? String(paymentFlatEntity.programNumber).concat(" - ", paymentFlatEntity.programName)
                 : String(paymentFlatEntity.programNumber),
@@ -15,5 +14,21 @@ export default class AmountsVsProgramRegionAdapter {
             montant: paymentFlatEntity.amount,
         };
         return amountsVsProgramRegionEntity;
+    }
+
+    static toDbo(entity: AmountsVsProgramRegionEntity): Omit<AmountsVsProgramRegionDbo, "_id"> {
+        return {
+            ...entity,
+        };
+    }
+
+    static toEntity(dbo: AmountsVsProgramRegionDbo): AmountsVsProgramRegionEntity {
+        return {
+            exerciceBudgetaire: dbo.exerciceBudgetaire,
+            montant: dbo.montant,
+            programme: dbo.programme,
+            regionAttachementComptable: dbo.regionAttachementComptable,
+            mission: dbo.mission,
+        };
     }
 }
